+++ R/client.R
@@ -594,6 +594,13 @@ OpenEOClient <- R6Class(
   
       query$req = req
       req = do.call(req_url_query,args = query)
+      
+      req = req_error(req,body=private$errorHandling)
+      
+      if (is.debugging()) {
+        req_dry_run(req)
+      }
+      
       response = req_perform(req)
       
       if (is.debugging()) {
@@ -615,7 +622,7 @@ OpenEOClient <- R6Class(
         
 
       } else {
-        private$errorHandling(response,url)
+        private$errorHandling(response)
       }
     },
     DELETE = function(endpoint,authorized=FALSE,query = list(),...) {
@@ -632,6 +639,13 @@ OpenEOClient <- R6Class(
 
       query$req = req
       req = do.call(req_url_query,args = query)
+      
+      req = req_error(req,body=private$errorHandling)
+      
+      if (is.debugging()) {
+        req_dry_run(req)
+      }
+      
       response = req_perform(req)
       
       # response = DELETE(url=url, config = header, ...)
@@ -648,14 +662,13 @@ OpenEOClient <- R6Class(
           return(TRUE)
         }
       } else {
-        private$errorHandling(response,url)
+        private$errorHandling(response)
       }
       
       
     },
     POST = function(endpoint,authorized=FALSE,data=list(),encodeType = "json",query = list(), raw=FALSE,parsed=TRUE,...) {
       url = paste(private$host,endpoint,sep="/")
-      
       req = request(url)
       req = req_method(req, method="POST")
       
@@ -667,35 +680,6 @@ OpenEOClient <- R6Class(
       if (authorized && !is.null(private$auth_client)) {
         header = private$addAuthorization(header)
       }
-      
-      # if (length(data) > 0) {
-      #   if (! raw) {
-      #     if (is.character(data)) {
-      #       # data = fromJSON(data,simplifyDataFrame = FALSE)
-      #       if (encodeType == "json") {
-      #         encodeType = "raw"
-      #         header = append(header, list(`Content-Type` = "application/json"))
-      #       }
-      #     } else if (is.list(data)) {
-      #       
-      #       if (encodeType == "json") {
-      #         encodeType = "raw"
-      #         header = append(header, list(`Content-Type` = "application/json"))
-      #         
-      #         
-      #         # data = do.call(toJSON, args = list(x = data,
-      #         #                                    auto_unbox = TRUE,
-      #         #                                    ...))
-      #           
-      #       }
-      #       
-      #     } else {
-      #       stop("Cannot interpret data - not a list that can be transformed into json")
-      #     }
-      #   } else {
-      #     header = append(header, list(`Content-Type` = "application/octet-stream"))
-      #   }
-      # }
       
       req = do.call(req_headers,header)
       query$req = req
@@ -712,16 +696,14 @@ OpenEOClient <- R6Class(
         if (length(data) > 0) req = req_body_json(req = req,data = data,auto_unbox = TRUE, ...)
       }
       
+      req = req_error(req,body=private$errorHandling)
+      
+      if (is.debugging()) {
+        req_dry_run(req)
+      }
       
       response = req_perform(req)      
-      # response=POST(
-      #   url= url,
-      #   config = header,
-      #   query = query,
-      #   body = data,
-      #   encode = encodeType
-      # )
-      
+
       if (is.debugging()) {
         print(response)
       }
@@ -744,7 +726,7 @@ OpenEOClient <- R6Class(
         }
         
       } else {
-        private$errorHandling(response,url)
+        private$errorHandling(response)
       }
     },
     PUT = function(endpoint, authorized=FALSE, data=list(),encodeType = "json",query = list(), raw=FALSE,parsed=TRUE,...) {
@@ -759,29 +741,6 @@ OpenEOClient <- R6Class(
         header = private$addAuthorization(header)
       }
       
-
-      # create JSON and prepare to send graph as post body
-      # if (is.character(data)) {
-      #   if (encodeType == "json") {
-      #     encodeType = "raw"
-      #     header = append(header, add_headers(`Content-Type` = "application/json"))
-      #   }
-      # } else if (is.list(data)) {
-      #   
-      #   if (encodeType == "json") {
-      #     encodeType = "raw"
-      #     header = append(header, add_headers(`Content-Type` = "application/json"))
-      #     
-      #     
-      #     # data = do.call(toJSON, args = list(x = data,
-      #     #                                    auto_unbox = TRUE,
-      #     #                                    ...))
-      #     
-      #   }
-      #   
-      # } else {
-      #   stop("Cannot interpret data - not a list that can be transformed into JSON")
-      # }
       req = do.call(req_headers,header)
       query$req = req
       req = do.call(req_url_query,args = query)
@@ -797,18 +756,13 @@ OpenEOClient <- R6Class(
         if (length(data) > 0) req = req_body_json(req = req,data = data,auto_unbox = TRUE, ...)
       }
       
+      req = req_error(req,body=private$errorHandling)
       
-      
+      if (is.debugging()) {
+        req_dry_run(req)
+      }
       
       response = req_perform(req)  
-
-      # response=PUT(
-      #   url= url,
-      #   config = header,
-      #   query = query,
-      #   body = data,
-      #   encode = encodeType
-      # )
       
       if (is.debugging()) {
         print(response)
@@ -826,7 +780,7 @@ OpenEOClient <- R6Class(
         okMessage = resp_body_json(response)
         return(okMessage)
       } else {
-        private$errorHandling(response,url)
+        private$errorHandling(response)
       }
     },
     PATCH = function(endpoint, authorized=FALSE, data=NULL, encodeType = NULL, parsed=TRUE, ...) {
@@ -840,20 +794,18 @@ OpenEOClient <- R6Class(
         req = do.call(req_headers,header)
       }
       
-      
-      
       params = list(url=url, 
                     config = header)
       
       if (!is.null(data)) {
-        # params = append(params, list(body = data))
         req = req_body_json(req = req,data = data,auto_unbox = TRUE, ...)
       }
       
-      # if (!is.null(encodeType)) {
-      #   params = append(params, list(encode = encodeType))
-      # }
-      # response = do.call("PATCH", args = params)
+      req = req_error(req,body=private$errorHandling)
+      
+      if (is.debugging()) {
+        req_dry_run(req)
+      }
       
       response = req_perform(req)
       
@@ -873,7 +825,7 @@ OpenEOClient <- R6Class(
         okMessage = resp_body_json(response)
         return(okMessage)
       } else {
-        private$errorHandling(response,url)
+        private$errorHandling(response)
       }
     },
 
@@ -901,19 +853,19 @@ OpenEOClient <- R6Class(
 
       return(header)
     },
-    errorHandling = function(response,url) {
-      if (class(response) == "response") {
+    errorHandling = function(response) {
+      if (class(response) == "httr2_response" || class(response) == "response") {
         # errorMessage = content(response)
         errorMessage = resp_body_json(response)
         if (!is.null(errorMessage[["message"]])) {
-          stop(paste("SERVER-ERROR:", errorMessage[["message"]]))
+          paste("SERVER-ERROR:", errorMessage[["message"]])
         } else {
           # if there is an uncaptured error from the server then just return it as is
-          stop(paste("SERVER-ERROR:", errorMessage))
+          paste("SERVER-ERROR:", errorMessage)
         }
       } else {
         # never happens? it is something else than response object
-        stop(response)
+        response
       }
     }
   )