+++ R/jobs.R
@@ -45,27 +45,34 @@ list_jobs = function(con=NULL) {
 #' the server remains open. This function allows to debug the code and check the results immediately. 
 #' Please keep in mind, that computational functions might be related to monetary costs, if no 'free' plan is available. 
 #' Make sure to keep the data selection relatively small, also some openEO service provider might offer limited processes support,
-#' e.g. not supporting UDFs at this endpoint.
+#' e.g. not supporting UDFs at this endpoint. When a file format is set, then the process graph will be parsed and the arguments for 
+#' 'save_result' will be replaced. If the 'stars' package is installed and parameter \code{as_stars} is set to TRUE, then the downloaded
+#' data is opened and interpreted into a stars object.
 #'
 #' @param graph a \code{\link{Graph}}, a function returning a \code{\link{ProcessNode}} as an endpoint or the \code{\link{ProcessNode}} 
 #' will return the results
 #' @param output_file storage location for the returned data
 #' @param budget numeric, maximum spendable amount for testing
 #' @param plan character, selection of a service plan
+#' @param as_stars logical to indicate if the data shall be interpreted as a stars object
+#' @param format character or \code{FileFormat} specifying the File format for the output, if 'save_result' is not
+#' set in the process then it will be added otherwise the value stated here will replace the original value
 #' @param con connected and authenticated openEO client (optional) otherwise \code{\link{active_connection}}
 #' is used.
-#' @param ... additional parameters passed to jsonlite::toJSON() (like 'digits')
+#' @param ... additional parameters passed to jsonlite::toJSON() (like 'digits') or additional arguments that shall 
+#' be passed to the openEO process 'save_result'
 #' 
-#' @return A connection to a file if the parameter 'output_file' was provided otherwise the raw binary data
+#' @return a local path to the downloaded file or a \code{stars} object if \code{as_stars=TRUE}
 #' 
 #' @importFrom methods as
 #' @export
-compute_result = function(graph, output_file = NULL, budget=NULL, plan=NULL, con=NULL, ...) {
+compute_result = function(graph, output_file = NULL, budget=NULL, plan=NULL, as_stars=FALSE, format = NULL, con=NULL, ...) {
     tryCatch({
         con = .assure_connection(con)
         
         output = list()
         
+        #TODO what if graph is already a Process?
         if (is.null(graph)) 
             stop("No process graph was defined. Please provide a process graph.")
         
@@ -82,6 +89,39 @@ compute_result = function(graph, output_file = NULL, budget=NULL, plan=NULL, con
         process = Process$new(id=NA,description = NA,
                               summary = NA,process_graph=graph)
         
+        # if format is set check if save_result is set, if not do that with the format stated, if it is 
+        # check if the formats match else replace
+        # more or less replace save_result of the graph with the customization stated in this function
+        if (length(format) > 0) {
+          save_node = .find_process_by_name(process,"save_result")
+          p = processes()
+          
+          dots = list(...)
+          
+          call_args = list(format = format)
+          
+          arg_names = names(formals(p$save_result))
+          save_result_dots = dots[which(arg_names %in% names(dots))]
+          
+          call_args = append(call_args, save_result_dots)
+          
+          if (length(save_node) == 0) {
+            
+            # not existent
+            call_args$data = process$getProcessGraph()$getFinalNode()
+            
+            # check for potentially multiple end nodes
+            
+          } else {
+            # check only first (or look for the final node)
+            saved_graph = save_node[[1]]
+            call_args$data = saved_graph$parameters$data
+          }
+          
+          saved_graph = do.call(p$save_result,call_args)
+          process = as(saved_graph,"Process")
+        }
+        
         job = list(
             process = process$serialize()
         )
@@ -94,22 +134,49 @@ compute_result = function(graph, output_file = NULL, budget=NULL, plan=NULL, con
             job$plan = plan
         }
         
+        is_tempfile = ifelse(length(output_file) == 0,TRUE,FALSE)
+        
         tag = "execute_sync"
-        # res = con$request(tag = tag, authorized = TRUE, data = job, encodeType = "json", raw = TRUE, ...)
         res = con$request(tag = tag, authorized = TRUE, data = job, encodeType = "json", parsed=FALSE, ...)
         
-        if (!is.null(output_file)) {
-            tryCatch({
-                message("Task result was sucessfully stored.")
-                writeBin(resp_body_raw(res), output_file)
-            }, error = function(err) {
-                stop(err)
-            })
-            
-            return(output_file)
-        } else {
-            return(resp_body_raw(res))
+        
+        
+        if (length(format) > 0 && length(output_file) == 0) {
+          if (is.character(format)) {
+            driver = format
+          } else if ("FileFormat" %in% class(format)) {
+            driver = format$name
+          } else {
+            message("Cannot derive a file extension. Using none which might affect data interpretation.")
+            driver = ""
+          }
+          
+          suffix = switch(driver, netCDF=".nc",GTiff=".tif",JSON=".json",default="")
+          output_file = tempfile(fileext = suffix)
         }
+        
+        tryCatch({
+          if (!dir.exists(dirname(output_file))) {
+            dir.create(dirname(output_file), recursive = TRUE)
+          }
+          writeBin(resp_body_raw(res), output_file)
+        }, error = function(err) {
+          stop(err)
+        })
+        
+        if (isTRUE(as_stars) && .is_package_installed("stars")) {
+          
+          obj=stars::read_stars(output_file,proxy=FALSE,quiet=TRUE)
+          
+          tryCatch({
+            return(obj)
+          }, finally={
+            if (is_tempfile) unlink(output_file)
+          })
+        } else {
+          return(output_file)
+        }
+        
     }, error = .capturedErrorToMessage)
 }
 
@@ -184,7 +251,7 @@ create_job = function(graph = NULL, title = NULL, description = NULL, plan = NUL
 #' @param con connected and authenticated openEO client (optional) otherwise \code{\link{active_connection}}
 #' is used.
 #' 
-#' @return the job_id of the defined job
+#' @return the job object of the now started job
 #' @export 
 start_job = function(job, log=FALSE, con=NULL) {
     tryCatch({
@@ -209,7 +276,7 @@ start_job = function(job, log=FALSE, con=NULL) {
             logs(job_id=job_id,con=con)
         }
         
-        invisible(success)
+        invisible(describe_job(job,con = con))
     }, error = .capturedErrorToMessage)
 }
 
@@ -319,24 +386,41 @@ list_results = function(job, con=NULL) {
 #' The function will fetch the results of a asynchronous job and will download all files stated in the links. The parameter
 #' 'folder' is the target location on the local computer.
 #' 
-#' @param job job object or the job_id for which the results are fetched
+#' @param job job object or the job_id for which the results are fetched. Also the return value of 
+#' \code{\link{list_results}} or its 'assets' field is also accepted.
 #' @param folder a character string that is the target path on the local computer
 #' @param con a connected and authenticated openEO connection (optional) otherwise \code{\link{active_connection}}
 #' is used.
 #' 
-#' @return a list of the target file paths
+#' @return a list of the target file paths or NULL if 'job' was incorrect
 #' 
 #' @importFrom utils download.file
 #' @export
 download_results = function(job, folder, con=NULL) {
     con = .assure_connection(con)
     
+    if (length(job) != 1) {
+      message("Parameter 'job' is not set.")
+      return(invisible(NULL))
+    }
+    
     if (!dir.exists(folder)) 
         dir.create(folder, recursive = TRUE)
-    results = list_results(con=con, job=job)
     
-    target_files = lapply(names(results$assets), function(file_name) {
-        link = results$assets[[file_name]]
+    if ("Job" %in% class(job) || is.character(job)) {
+      results = list_results(con=con, job=job)
+      assets = results$assets
+    } else if ("AssetList" %in% class(job)) {
+      assets = job
+    } else if ("ResultList" %in% class(job)) {
+      assets = job$assets
+    } else {
+      message("Parameter 'job' was not interpretable.")
+      return(invisible(NULL))
+    }
+    
+    target_files = lapply(names(assets), function(file_name) {
+        link = assets[[file_name]]
         
         href = link$href
         type = link$type
@@ -346,7 +430,11 @@ download_results = function(job, folder, con=NULL) {
         
         file_path = paste0(folder, file_name)
         
-        download.file(href, file_path, mode = "wb")
+        req = request(href)
+        req = req_method(req, method="GET")
+        response = req_perform(req)
+        
+        writeBin(resp_body_raw(response), file_path)
         
         return(file_path)
     })